loader.py
from main import main
from api import BHYG
from bilibili_util import BilibiliClient
from push import do_push

import sys
import PIL
import httpx
import loguru
import prompt_toolkit
import qrcode
import questionary
import cryptography
import httpx
import sentry_sdk

from cryptography.hazmat.backends import default_backend
from cryptography.hazmat.primitives import serialization
from cryptography.hazmat.primitives.asymmetric import rsa, padding
from cryptography.hazmat.primitives import hashes

# try:
#     import bili_ticket_gt_python
# except:
#     pass

import notifypy

if sys.platform == "win32":
    import wmi

import playsound3

from Crypto.Cipher import AES
from Crypto.Util.Padding import pad, unpad
from sentry_sdk.scrubber import EventScrubber
from sentry_sdk.integrations.loguru import LoguruIntegration, LoggingLevels
import sentry_sdk
from bilibili_util import BilibiliClient

if __name__ == "__main__":
    main()
