requirements.txt
Nuitka
httpx>=0.26.0
loguru>=0.7.2
prompt_toolkit>=3.0.43
qrcode>=7.4.2
questionary>=2.0.1
cryptography>=41.0.5
httpx[http2,brotli,zstd,socks]>=0.28.0
sentry-sdk>=2.27.0
Cython
setuptools
pyinstaller
pillow
notify-py
pycryptodome
wmi
cryptography
playsound3
py-machineid